authentication and tokens are added

+++ controllers/expense.js
@@ -3,8 +3,11 @@ const Expense= require('../models/expense');
 
 exports.postExpense= async(req,res,next)=>{
     try {
-        const {amount, description, category } = req.body;
-        const data = await Expense.create({ amount: amount, description:description, category:category });
+        const amount=req.body.amount
+        const description=req.body.description
+        const category=req.body.category
+        const data = await  req.user.createExpense({ amount: amount, description:description, category:category});
+        //Expense.create({ amount: amount, description:description, category:category,userId:req.user.id });
         res.status(201).json({ newUserExpense: data });
       } catch (err) {
         console.error(err);
@@ -14,7 +17,7 @@ exports.postExpense= async(req,res,next)=>{
 
  exports.getExpense= async(req,res,next)=>{
     try{
-        const expenses= await Expense.findAll();
+        const expenses= await  req.user.getExpenses();               //Expense.findAll({where :{userId: req.user.id}});
         res.status(200).json({allExpense: expenses});
       }
       catch(err){
@@ -25,12 +28,12 @@ exports.postExpense= async(req,res,next)=>{
 
 exports.deleteExpense= async(req,res,next)=>{
     const id = req.params.id;
-    //console.log(id);
+    console.log(id);
     try {
       if(!req.params.id){
         return res.status(400).json({error: 'Id is missing'})
       }
-      await Expense.destroy({where: { id: id}});
+      await Expense.destroy({where: { id: id , userId: req.user.id}});
       res.sendStatus(200);
     } catch (error) {
       console.error(error);
